refactor collections and initialGamestate, rename fields/collections, add numPlayers and difficultyLevel as parameters
