+++ src/utils/index.js
@@ -1,0 +1,5 @@
+export { default as cities } from './initialCities'
+export { default as unusedCityCards } from './unusedCityCards'
+export { default as unusedInfectionCards } from './unusedInfectionCards'
+export { default as collections } from './collections'
+export { default as init } from './initialGameState'
