+++ src/utils/unusedCityCards.js
@@ -1,0 +1,196 @@
+const unusedCityCards = {
+  Algiers: {
+    name: "Algiers",
+    color: "black",
+  },
+  Atlanta: {
+    name: "Atlanta",
+    color: "blue"
+  },
+  Baghdad: {
+    name: "Baghdad",
+    color: "black"
+  },
+  Bangkok: {
+    name: "Bangkok",
+    color: "red"
+  },
+  Beijing: {
+    name: "Beijing",
+    color: "red"
+  },
+  Bogota: {
+    name: "Bogota",
+    color: "yellow"
+  },
+  BuenosAires: {
+    name: "Buenos Aries",
+    color: "yellow"
+  },
+  Cairo: {
+    name: "Cairo",
+    color: "black"
+  },
+  Chennai: {
+    name: "Chennai",
+    color: "black"
+  },
+  Chicago: {
+    name: "Chicago",
+    color: "blue"
+  },
+  Delhi: {
+    name: "Delhi",
+    color: "black"
+  },
+  Essen: {
+    name: "Essen",
+    color: "blue"
+  },
+  HoChiMinhCity: {
+    name: "Ho Chi Minh City",
+    color: "red"
+  },
+  HongKong: {
+    name: "Hong Kong",
+    color: "red"
+  },
+  Istanbul: {
+    name: "Istanbul",
+    color: "black"
+  },
+  Jakarta: {
+    name: "Jakarta",
+    color: "red"
+  },
+  Johannesburg: {
+    name: "Johannesburg",
+    color: "yellow"
+  },
+  Karachi: {
+    name: "Karachi",
+    color: "black"
+  },
+  Khartoum: {
+    name: "Khartoum",
+    color: "yellow"
+  },
+  Kinshasa: {
+    name: "Kinshasa",
+    color: "yellow"
+  },
+  Kolkata: {
+    name: "Kolkata",
+    color: "black"
+  },
+  Lagos: {
+    name: "Lagos",
+    color: "yellow"
+  },
+  Lima: {
+    name: "Lima",
+    color: "yellow"
+  },
+  London: {
+    name: "London",
+    color: "blue"
+  },
+  LosAngeles: {
+    name: "Los Angeles",
+    color: "yellow"
+  },
+  Madrid: {
+    name: "Madrid",
+    color: "blue"
+  },
+  Manila: {
+    name: "Manila",
+    color: "red"
+  },
+  MexicoCity: {
+    name: "Mexico City",
+    color: "yellow"
+  },
+  Miami: {
+    name: "Miami",
+    color: "yellow"
+  },
+  Milan: {
+    name: "Milan",
+    color: "blue"
+  },
+  Montreal: {
+    name: "Montreal",
+    color: "blue"
+  },
+  Moscow: {
+    name: "Moscow",
+    color: "black"
+  },
+  Mumbai: {
+    name: "Mumbai",
+    color: "black"
+  },
+  NewYork: {
+    name: "New York",
+    color: "blue"
+  },
+  Osaka: {
+    name: "Osaka",
+    color: "red"
+  },
+  Paris: {
+    name: "Paris",
+    color: "blue"
+  },
+  Riyadh: {
+    name: "Riyadh",
+    color: "black"
+  },
+  SanFrancisco: {
+    name: "San Francisco",
+    color: "blue"
+  },
+  Santiago: {
+    name: "Santiago",
+    color: "yellow"
+  },
+  SaoPaulo: {
+    name: "Sao Paulo",
+    color: "yellow"
+  },
+  Seoul: {
+    name: "Seoul",
+    color: "red"
+  },
+  Shanghai: {
+    name: "Shanghai",
+    color: "red"
+  },
+  StPetersburg: {
+    name: "St. Petersburg",
+    color: "black"
+  },
+  Sydney: {
+    name: "Sydney",
+    color: "red"
+  },
+  Taipei: {
+    name: "Taipei",
+    color: "red"
+  },
+  Tehran: {
+    name: "Tehran",
+    color: "black"
+  },
+  Tokyo: {
+    name: "Tokyo",
+    color: "red"
+  },
+  Washington: {
+    name: "Washington",
+    color: "blue"
+  }
+}
+
+export default unusedCityCards
